Move downloadFile -> utils/download
